Split the DRPAI class into 3 classes, controller, connection, yolo
